ai_app/requirements.txt
# AI Application Dependencies
tensorflow==2.15.0
pytesseract==0.3.10
opencv-python==4.9.0.80
pdf2image==1.17.0
Pillow==10.2.0
numpy==1.26.4
pandas==2.2.0

# API Dependencies
fastapi==0.109.2
python-multipart==0.0.9
uvicorn==0.27.1
pydantic==2.6.1

# Testing and Utils
pytest==8.0.2
python-dotenv==1.0.1 